cleanup clustering
