+++ visMOP/python_scripts/cluster_layout.py
@@ -367,22 +367,6 @@ class Cluster_layout:
             node_num / total_num_nodes for node_num in self.cluster_node_nums
         ]
 
-        # set AN_RATIO and MIN_SIDE_RATIO depending on number of Clusters
-        self.AN_RATIO = 0.32 if self.num_cluster > 10 else 0.2
-        self.MIN_SIDE_T = 0.12 if self.num_cluster > 10 else 0.2
-        self.num_best_dist_loops = 500 if self.num_cluster > 10 else 1500
-        self.num_cost_min_loops = 0 if self.num_cluster > 10 else 5
-
-        # for each cluster idenify all positions where rel. distances to it are given
-        self.p_f_m = defaultdict(list)
-        for i in range(self.num_cluster):
-            pos_m = []
-            for j in range(i + 1, self.num_cluster):
-                pos = self.num_cluster * i + j - ((i + 2) * (i + 1)) // 2
-                pos_m.append(pos)
-                self.p_f_m[j].append(pos)
-            self.p_f_m[i] += pos_m
-
         initial_clusters_center = self.get_cluster_centers(self.initial_node_pos)
         self.clusters_center = initial_clusters_center
         self.weights = self.getClustersWeights()
@@ -391,288 +375,12 @@ class Cluster_layout:
             self.get_cluster_centers(self.positions_dict_clustering)
         )
         print("Clusters identified")
-        # outcommented for test of voronoi layout
-        # self.area_num_node_ratio_Pool()
         print(cluster_length)
-        print("Relative distances between clusters and cluster weights calculated")
-        # outcommented for test of voronoi layout
-        # self.final_node_pos = self.getNodePositions(pathways_root_names)
-        print("final node positions identified")
         endTime = time.time()
         # self.get_stats()
         # self.get_stat_plots()
         print(
             "Time for Cluster Layout calculation {:.3f} s".format((endTime - startTime))
-        )
-
-    def area_num_node_ratio_Pool(self):
-        self.min_nn_ratios = [
-            nn_r - self.AN_RATIO * nn_r for nn_r in self.node_num_ratio
-        ]
-        try:
-            np.random.seed()
-            self.clusters_area, initial_area_side_ratio = self.getSizeOfClustersRegion(
-                self.clusters_center
-            )
-
-            nn_a_comp = [
-                a_r >= (nn_r - self.AN_RATIO * nn_r) and min_side >= self.MIN_SIDE_T
-                for nn_r, [a_r, min_side] in zip(
-                    self.node_num_ratio, initial_area_side_ratio
-                )
-            ]
-            found_ideal = True if sum(nn_a_comp) == len(nn_a_comp) else False
-            (
-                init_area_comp,
-                num_area_w_greater_diff,
-                init_min_side_comp,
-                num_side_smaller,
-            ) = self.calc_com_values(self.min_nn_ratios, initial_area_side_ratio, 0)
-            initial_cost = self.evaluteCostFunction(self.clusters_center)
-        except:
-            found_ideal = False
-            init_area_comp, init_min_side_comp, initial_cost = inf, inf, inf
-            num_area_w_greater_diff, num_side_smaller = (
-                self.num_cluster,
-                self.num_cluster,
-            )
-        if not found_ideal:
-            pool = Pool(self.pool_size)
-            self.MAX_RUNS_PER_PROCESS = 15
-            multiple_results = []
-            simplified_func = partial(
-                self.area_num_node_ratio_optFunc,
-                area_comp=init_area_comp,
-                num_area_w_greater_diff=num_area_w_greater_diff,
-                num_side_smaller=num_side_smaller,
-                diff_min_side_to_success=init_min_side_comp,
-                clusters_area=self.clusters_area,
-                initial_cost=initial_cost,
-            )
-            for result in pool.imap_unordered(
-                simplified_func, [self.clusters_center] * self.pool_size
-            ):
-                if result:  # first to succeed:
-                    if result[0]:
-                        print("POOL OPTI DONE ", result[3])
-                        found_ideal = True
-                        self.clusters_area = result[3]
-                        smallest_area_diff = result[1]
-                        smalles_min_side_diff = result[2]
-                        C_cost_best = result[4]
-                        break
-                    else:
-                        multiple_results.append(result)
-            pool.terminate()  # kill all remaining tasks
-            # Wait for all processes to end:
-            pool.join()
-            # success, area_comp, diff_min_side_to_success, clusters_area
-        # with open('cluster_areas_03.pkl', "rb") as f:
-        #     self.clusters_area = pickle.load(f)
-
-        if not found_ideal:
-            smallest_area_diff = np.Infinity
-            smalles_min_side_diff = np.Infinity
-            C_cost_best = np.Infinity
-            for result in multiple_results:
-                if (result[1] < smallest_area_diff) or (
-                    (
-                        result[1] == smallest_area_diff - result[1] <= 0.01
-                        or (abs(result[1] - smallest_area_diff) < 0.01)
-                    )
-                    and result[2] < smalles_min_side_diff
-                ):
-                    smallest_area_diff = result[1]
-                    smalles_min_side_diff = result[2]
-                    self.clusters_area = result[3]
-                    C_cost_best = result[4]
-            print("suboptimal solution found: ")
-            print("area_side_ratio", self.clusters_area)
-            print(
-                "min_side/max side",
-                [get_min_max_side(area, self.l_max) for area in self.clusters_area],
-            )
-            print("node_num_ratio", self.node_num_ratio)
-
-        print("Cluster sizes identified")
-        print("smallest_area_diff:", init_area_comp)
-        print("smalles_min_side_diff: ", init_min_side_comp)
-        print("innitial_rel_dist", self.relative_distances)
-        print("initial_cost: ", initial_cost)
-        print("self.node_num_ratio", self.node_num_ratio)
-        print("initial_area_side_ratio", initial_area_side_ratio)
-        print(
-            "min_side",
-            [get_area_size(ma, True, self.l_max)[1] for ma in self.clusters_area],
-        )
-
-    def area_num_node_ratio_optFunc(
-        self,
-        initial_clusters_center,
-        area_comp,
-        num_area_w_greater_diff,
-        num_side_smaller,
-        diff_min_side_to_success,
-        clusters_area,
-        initial_cost,
-    ):
-        np.random.seed()
-        runs = 0
-        total_num_runs = 0
-        success = False
-        clusters_center = initial_clusters_center
-        runtime_threshold_MS = 0
-        prev_best_cost_C = initial_cost
-        min_nn_ratios = self.min_nn_ratios
-        found_better = 0
-        while not success and total_num_runs <= self.MAX_RUNS_PER_PROCESS:
-            best_cost_C, cluster_center = self.getClusterPos(
-                clusters_center, prev_best_cost_C
-            )
-            try:
-                new_clusters_area, new_area_side_ratio = self.getSizeOfClustersRegion(
-                    cluster_center
-                )
-                nn_a_comp = [
-                    a_r >= min_nn_ratio
-                    and min_side >= (self.MIN_SIDE_T - runtime_threshold_MS)
-                    for min_nn_ratio, (a_r, min_side) in zip(
-                        min_nn_ratios, new_area_side_ratio
-                    )
-                ]
-
-                success = True if sum(nn_a_comp) == len(nn_a_comp) else False
-                if success:
-                    clusters_area = new_clusters_area
-                    area_comp = 0
-                    diff_min_side_to_success = 0
-                elif runs == self.MAX_RUNS_PER_PROCESS / 2:
-                    runs = 0
-                    # start at random new position
-                    cluster_center = np.random.uniform(0, 1, (self.num_cluster, 2))
-                    prev_best_cost_C = None
-                else:
-                    (
-                        new_area_comp,
-                        new_num_area_w_greater_diff,
-                        new_diff_min_side_to_success,
-                        new_num_side_smaller,
-                    ) = self.calc_com_values(
-                        min_nn_ratios, new_area_side_ratio, runtime_threshold_MS
-                    )
-                    area_better = (
-                        new_area_comp <= area_comp
-                        and new_num_area_w_greater_diff <= num_area_w_greater_diff
-                    )
-                    area_sim = (
-                        new_diff_min_side_to_success <= diff_min_side_to_success
-                        or new_area_comp <= 0.01
-                    ) and (new_num_area_w_greater_diff <= num_area_w_greater_diff)
-                    min_side_better = (
-                        area_sim and new_num_side_smaller <= num_side_smaller
-                    )
-                    max_side_not_whole_side = (
-                        sum(
-                            [
-                                not (max_s < 0.99 or max_s > 0.99 and min_s > 0.25)
-                                for min_s, max_s in [
-                                    get_min_max_side(area, self.l_max)
-                                    for area in new_clusters_area
-                                ]
-                            ]
-                        )
-                        == 0
-                    )
-                    if (
-                        area_better
-                        or (
-                            (abs(new_area_comp - area_comp) <= 0.01) and min_side_better
-                        )
-                    ) and max_side_not_whole_side:  #  and best_cost_C < prev_best_cost_C):
-                        found_better += 1
-                        area_comp = new_area_comp
-                        diff_min_side_to_success = new_diff_min_side_to_success
-                        clusters_area = new_clusters_area
-                        num_area_w_greater_diff = new_num_area_w_greater_diff
-                        num_side_smaller = new_num_side_smaller
-                        prev_best_cost_C = best_cost_C
-                    # else:
-                    #     if max_side_not_whole_side and not area_better and (abs(new_area_comp - area_comp) <= 0.01) and new_num_side_smaller <= num_side_smaller:
-                    #         print(area_sim)
-
-                    runs += 1
-                if total_num_runs == self.MAX_RUNS_PER_PROCESS / 2:
-                    min_nn_ratios = [
-                        nn_r - (self.AN_RATIO + 0.05) * nn_r
-                        for nn_r in self.node_num_ratio
-                    ]
-                    runtime_threshold_MS = 0.01
-                total_num_runs += 1
-
-            except IndexError:
-                cluster_pairs = combinations(range(len(clusters_center)), 2)
-                cluster_dists = {
-                    m_p: sum(
-                        [
-                            abs(clusters_center[m_p[0]][0] - clusters_center[m_p[1]][0])
-                            > 0.0001,
-                            abs(clusters_center[m_p[0]][1] - clusters_center[m_p[1]][1])
-                            > 0.0001,
-                        ]
-                    )
-                    for m_p in cluster_pairs
-                }
-                none_correct_dists = list(
-                    sum(
-                        [
-                            m_p
-                            for m_p, corr_dist in cluster_dists.items()
-                            if corr_dist < 2
-                        ],
-                        (),
-                    )
-                )
-                # makes sure that there is enough space too find a area for each cluster in divideSpaceForTwoClusters()
-
-                print(
-                    "no possible area divide found with number of none correct distances",
-                    none_correct_dists,
-                )
-                pass
-
-        print("C_cost =", self.evaluteCostFunction(cluster_center))
-        print("better Solutions found", found_better)
-
-        return (
-            success,
-            area_comp,
-            diff_min_side_to_success,
-            clusters_area,
-            self.evaluteCostFunction(cluster_center),
-        )
-
-    def calc_com_values(self, min_nn_ratios, new_area_side_ratio, runtime_threshold_MS):
-        area_diff_to_success = [
-            min_nn_ratio - a_r
-            for min_nn_ratio, (a_r, _) in zip(min_nn_ratios, new_area_side_ratio)
-            if min_nn_ratio > a_r
-        ]
-        new_area_comp = sum(area_diff_to_success)
-        new_num_area_w_greater_diff = sum(
-            [diff >= 0.08 for diff in area_diff_to_success]
-        )
-        min_side_diff_to_success = [
-            (self.MIN_SIDE_T - runtime_threshold_MS) - min_side
-            for _, min_side in new_area_side_ratio
-            if (self.MIN_SIDE_T - runtime_threshold_MS) > min_side
-        ]
-        new_diff_min_side_to_success = sum(min_side_diff_to_success)
-        new_num_side_smaller = sum([diff >= 0.02 for diff in min_side_diff_to_success])
-        return (
-            new_area_comp,
-            new_num_area_w_greater_diff,
-            new_diff_min_side_to_success,
-            new_num_side_smaller,
         )
 
     # default values per omic and pos
@@ -882,378 +590,3 @@ class Cluster_layout:
         relative_distances = [dist / max_distance for dist in all_distances]
 
         return relative_distances
-
-    def evaluateRelativDistanceSimilarity(self, original, new, dist_sim_threshold=0.1):
-        """evaluate if clusters have the same realtive distance as in the original clustering layout"""
-
-        distance_similarities = [
-            (rd_org - dist_sim_threshold) <= rd_new <= (rd_org + dist_sim_threshold)
-            for rd_org, rd_new in zip(original, new)
-        ]
-
-        distance_similarities = [
-            ds if m not in self.p_f_m[0] else True
-            for m, ds in enumerate(distance_similarities)
-        ]
-
-        dist_sim_sum = sum(distance_similarities)
-        possibleLayout = dist_sim_sum == len(distance_similarities)
-        return possibleLayout, distance_similarities
-
-    def evaluteCostFunction(self, cur_cluster_centers):
-        """
-        Parameter: Weight of cluster h and k w(h,k), Manhattan distance between cluster h and k d(h,k)
-        C = sum over (w(h,k) * d(h,k))
-        """
-        num_clusters = len(self.clusters)
-        distances = np.zeros((num_clusters, num_clusters))
-        for cluster_pair in combinations(range(len(self.clusters)), 2):
-            distances[cluster_pair[0], cluster_pair[1]] = distance.cityblock(
-                cur_cluster_centers[cluster_pair[0]],
-                cur_cluster_centers[cluster_pair[1]],
-            )
-
-        C = sum(
-            sum([np.array(w) * np.array(d) for w, d in zip(self.weights, distances)])
-        )
-
-        return C
-
-    def getClusterPos(self, clusters_center, prev_best_C):
-        """
-        1. getClustersWeights
-        2. getSizeOfClustersRegion
-        3. initial cluster layout R given bei umap or pca
-        4. repeat x times
-            4.1 repeat until a positioning found which satifies the condition that clusters don't overlap and relative distances are the same
-                4.1.1 randomly move clusters R1
-            4.2 partly optimize R1
-            4.3 R = R1 if C1 < C
-        5. return R
-        """
-        internal_clusters_center = clusters_center
-        best_C = (
-            self.evaluteCostFunction(internal_clusters_center)
-            if prev_best_C is None
-            else prev_best_C
-        )
-        iter_num = 0
-        possibleLayout_found = False
-
-        while iter_num <= self.num_cost_min_loops and not possibleLayout_found:
-            new_cluster_centers = self.find_layout_with_best_possible_rel_dists(
-                internal_clusters_center
-            )
-            new_C = self.evaluteCostFunction(new_cluster_centers)
-            if new_C < best_C or (
-                iter_num == self.num_cost_min_loops and not possibleLayout_found
-            ):
-                possibleLayout_found = True
-                internal_clusters_center = new_cluster_centers
-                best_C = new_C
-            iter_num += 1
-        return best_C, internal_clusters_center
-
-    def find_layout_with_best_possible_rel_dists(self, clusters_center):
-        distance_similarities = [0] * len(self.relative_distances)
-        cur_cluster_center = clusters_center
-        run = 0
-        found_better_possible_center = False
-        found_perfect_center = False
-        while run <= self.num_best_dist_loops and not found_perfect_center:
-            new_cluster_centers = [
-                self.getNewClusterCenter(
-                    cluster_center,
-                    mn,
-                    sum(distance_similarities[p] for p in self.p_f_m[mn]),
-                )
-                for mn, cluster_center in enumerate(cur_cluster_center)
-            ]
-            new_rel_distances = self.getRealtiveDistancesBetweenClusters(
-                new_cluster_centers
-            )
-
-            (
-                found_perfect_center,
-                distance_similarities_new,
-            ) = self.evaluateRelativDistanceSimilarity(
-                self.relative_distances, new_rel_distances, self.dist_sim_threshold
-            )
-            if sum(distance_similarities_new) > sum(distance_similarities):
-                distance_similarities = distance_similarities_new
-                cur_cluster_center = new_cluster_centers
-                found_better_possible_center = True
-            elif run == self.num_best_dist_loops and not found_better_possible_center:
-                cur_cluster_center = new_cluster_centers
-            run += 1
-
-        return cur_cluster_center
-
-    def getNewClusterCenter(
-        self, cluster_center, cluster_num, distance_similarities_value
-    ):
-        # je mehr die distancen stimmen desto weniger wahrscheinlich wird das cluster center erschoben
-        dist_bonus = (
-            distance_similarities_value / (self.num_cluster) - 0.2
-            if cluster_num != 0
-            else 0
-        )
-        change_pos = np.random.binomial(1, max(0.1, min(1, 0.1 + dist_bonus))) == 0
-        new_center = cluster_center
-        if change_pos:
-            # max movement in neg/ pos x direction: m_c[0] - 0.1 / 1-m_c[0] -0.1
-            for coord in [0, 1]:
-                new_center[coord] = np.random.uniform(0.1, 0.95)
-        return new_center
-
-    def getSizeOfClustersRegion(self, cluster_center):
-        """
-        calculate For dir for cluster
-        get max nodes in vertical and horizontal position
-        """
-
-        l_max = 20 * math.sqrt(max(self.cluster_node_nums))
-        self.l_max = l_max
-
-        new_centers = [
-            [cluster_center[0] * l_max, cluster_center[1] * l_max]
-            for cluster_center in cluster_center
-        ]
-
-        cluster_id_sorted_by_num_nodes = [
-            x
-            for _, x in sorted(
-                zip(self.cluster_node_nums, list(range(len(self.cluster_node_nums)))),
-                reverse=True,
-            )
-        ]
-
-        cluster_regions = self.divideSpaceForTwoClusters(
-            new_centers,
-            # cluster_x_y_distances,
-            [0, l_max, 0, l_max],
-            list(range(self.num_cluster)),
-            sum(self.cluster_node_nums),
-            cluster_id_sorted_by_num_nodes,
-        )
-        area_list = [[]] * self.num_cluster
-        for (area, cluster_num) in cluster_regions:
-            area_list[cluster_num] = area
-
-        final_area_size = [get_area_size(area, True, l_max) for area in area_list]
-        total_area = sum([size for size, _ in final_area_size])
-
-        area_side_ratio = [
-            [area / total_area, min_side] for area, min_side in final_area_size
-        ]
-
-        return area_list, area_side_ratio
-
-    def divideSpaceForTwoClusters(
-        self,
-        cluster_centers,
-        area_to_divide,
-        clusters_in_area,
-        total_sum_nodes_in_area,
-        cluster_id_sorted_by_num_nodes,
-    ):
-        if len(clusters_in_area) == 0:
-            print(
-                "############################################## I SHOULD NOT BE #####################################"
-            )
-            print(clusters_in_area)
-        if len(clusters_in_area) == 1:
-            return [[area_to_divide, clusters_in_area[0]]]
-
-        cluster_pair = []
-        for cluster in cluster_id_sorted_by_num_nodes:
-            if cluster in clusters_in_area:
-                cluster_pair.append(cluster)
-                if len(cluster_pair) == 2:
-                    break
-
-        cluster_1 = cluster_pair[0]
-        cluster_2 = cluster_pair[1]
-
-        max_dist = max(
-            abs(
-                cluster_centers[cluster_pair[0]][0]
-                - cluster_centers[cluster_pair[1]][0]
-            ),
-            abs(
-                cluster_centers[cluster_pair[0]][1]
-                - cluster_centers[cluster_pair[1]][1]
-            ),
-        )
-        # max dist is on x axis
-        x_dist = cluster_centers[cluster_1][0] - cluster_centers[cluster_2][0]
-        y_dist = cluster_centers[cluster_1][1] - cluster_centers[cluster_2][1]
-        area_1 = deepcopy(area_to_divide)
-        area_2 = deepcopy(area_to_divide)
-        if 0.9 < abs(x_dist) / abs(y_dist) < 1.1:
-            max_dist = abs(x_dist) if np.random.binomial(1, 0.5) else abs(y_dist)
-        # change x coord
-        if max_dist == abs(x_dist):
-            cluster_1_area_pos_to_ad, cluster_2_area_pos_to_ad, min_border_pos = (
-                [1, 0, cluster_centers[cluster_1][0]]
-                if x_dist < 0
-                else [0, 1, cluster_centers[cluster_2][0]]
-            )
-        # change y coord
-        else:
-            y_dist = cluster_centers[cluster_1][1] - cluster_centers[cluster_2][1]
-            cluster_1_area_pos_to_ad, cluster_2_area_pos_to_ad, min_border_pos = (
-                [3, 2, cluster_centers[cluster_1][1]]
-                if y_dist < 0
-                else [2, 3, cluster_centers[cluster_2][1]]
-            )
-
-        (
-            area_1_best,
-            area_2_best,
-            clusters_in_area_1_best,
-            num_nodes_area_2,
-            num_nodes_area_1_best,
-            num_nodes_area_2_best,
-            score_min,
-        ) = [[], [], [], [], 0, 0, inf]
-        area_t_d_size = get_area_size(area_to_divide)
-        for abst in np.arange(0.0005, max_dist - 0.0005, 0.001):
-            area_1[cluster_1_area_pos_to_ad] = min_border_pos + abst
-            area_2[cluster_2_area_pos_to_ad] = min_border_pos + abst
-            clusters_in_area_1 = [
-                cluster
-                for cluster in clusters_in_area
-                if cluster_center_in_area(cluster_centers[cluster], area_1)
-            ]
-            num_nodes_area_1 = sum(
-                [self.cluster_node_nums[cluster] for cluster in clusters_in_area_1]
-            )
-            num_nodes_area_2 = total_sum_nodes_in_area - num_nodes_area_1
-            # score = abs(self.get_area_size(area_1)*num_nodes_area_2/total_sum_nodes_in_area -
-            #             self.get_area_size(area_2)*num_nodes_area_1/total_sum_nodes_in_area)
-
-            score = max(
-                0,
-                abs(
-                    get_area_size(area_1) / area_t_d_size
-                    - num_nodes_area_1 / total_sum_nodes_in_area
-                ),
-            ) + max(
-                0,
-                abs(
-                    get_area_size(area_2) / area_t_d_size
-                    - num_nodes_area_2 / total_sum_nodes_in_area
-                ),
-            )
-
-            if score < score_min:
-                score_min = score
-                clusters_in_area_1_best = deepcopy(clusters_in_area_1)
-                clusters_in_area_2_best = [
-                    cluster
-                    for cluster in clusters_in_area
-                    if cluster not in clusters_in_area_1_best
-                ]
-
-                area_1_best = deepcopy(area_1)
-                area_2_best = deepcopy(area_2)
-                num_nodes_area_1_best = deepcopy(num_nodes_area_1)
-                num_nodes_area_2_best = deepcopy(num_nodes_area_2)
-
-        return self.divideSpaceForTwoClusters(
-            cluster_centers,
-            # deepcopy(cluster_ia_x_y_distances),
-            area_1_best,
-            clusters_in_area_1_best,
-            num_nodes_area_1_best,
-            cluster_id_sorted_by_num_nodes,
-        ) + self.divideSpaceForTwoClusters(
-            cluster_centers,
-            # deepcopy(cluster_ia_x_y_distances),
-            area_2_best,
-            clusters_in_area_2_best,
-            num_nodes_area_2_best,
-            cluster_id_sorted_by_num_nodes,
-        )
-
-    def getNodePositions(self, pathways_root_names):
-        """
-        1. for all clusters
-            1.1 get force directed layout
-            1.2. adjust position to cluster Region coordinates
-        2. return node positions
-        """
-        node_positions = {}
-        total_num_nodes = len(list(self.data_table.index))
-        # with open('cluster_areas_03.pkl', "rb") as f:
-        #     self.clusters_area = pickle.load(f)
-        for cluster_num, (cluster, cluster_area) in enumerate(
-            zip(self.clusters, self.clusters_area)
-        ):
-            sub_graph = self.full_graph.subgraph(cluster)
-            print("cluster", cluster_area)
-            cluster_node_positions = get_adjusted_force_dir_node_pos(
-                sub_graph,
-                cluster_num,
-                pathways_root_names,
-                total_num_nodes,
-                self.l_max,
-                cluster_area,
-            )
-            adjusted_node_positions = normalize_2D_node_pos_in_range(
-                cluster_node_positions, cluster_area, True
-            )
-
-            node_positions = {**node_positions, **adjusted_node_positions}
-
-        return node_positions
-
-    def get_final_node_positions(self):
-        max_ext = 20
-        node_positions = normalize_2D_node_pos_in_range(
-            self.final_node_pos, [0, max_ext, 0, max_ext], True
-        )
-        cluster_dic = {
-            pathway: cluster_num
-            for cluster_num, cluster in enumerate(self.clusters)
-            for pathway in cluster
-        }
-        for root, subpathways in self.reactome_roots.items():
-            try:
-                maj_cluster_num = most_frequent(
-                    [
-                        cluster_dic[pathway]
-                        for pathway in subpathways
-                        if pathway in cluster_dic.keys()
-                    ]
-                )
-            except:
-                print(
-                    "{} Is a Root with associated measurements NOT in subpathways".format(
-                        root
-                    )
-                )
-                maj_cluster_num = -1
-            node_positions[root] = [0, 0, maj_cluster_num]
-
-        return node_positions
-
-    def get_cluster_areas(self):
-        max_ext = 20
-
-        flatten_list = list(map(list, zip(*self.clusters_area)))
-
-        min_x, max_x = [min(flatten_list[0]), max(flatten_list[1])]
-        min_y, max_y = [min(flatten_list[2]), max(flatten_list[3])]
-        norm_areas = [] if self.noise_cluster_exists else [[]]
-        for cluster in self.clusters_area:
-            norm_area = [
-                normalize_val_in_range(cluster[0], min_x, max_x, [0, max_ext]),
-                normalize_val_in_range(cluster[1], min_x, max_x, [0, max_ext]),
-                normalize_val_in_range(cluster[2], min_y, max_y, [0, max_ext]),
-                normalize_val_in_range(cluster[3], min_y, max_y, [0, max_ext]),
-            ]
-            norm_areas.append(norm_area)
-
-        return norm_areas
